new example to check destroing of controls
