+++ examples/widgets/showUiSpinbox.js
@@ -1,29 +1,19 @@
 /* eslint-disable unicorn/filename-case */
 
 const libui = require('../..');
-function createWindow() {
-	const win = new libui.UiWindow('UiSpinbox example', 320, 60, true);
 
-	win.margined = true;
+const win = new libui.UiWindow('UiSpinbox example', 320, 60, true);
+win.margined = true;
 
-	const widget = new libui.UiSpinbox();
-	widget.value = 42;
-	widget.onChanged(() => {
-		console.log(`value changed`);
-	});
-	win.setChild(widget);
+const widget = new libui.UiSpinbox();
+widget.value = 'sample text';
+win.setChild(widget);
 
-	win.onClosing(() => {
-		win.close();
-		global.gc();
-		libui.stopLoop();
-	});
+win.onClosing(() => {
+	win.close();
+	libui.stopLoop();
+});
 
-	return win;
-}
-
-const win = createWindow();
 win.show();
-global.gc();
 
 libui.startLoop();
